This changes the luminance calculation to be done once up front rather than repeatedly during the energy calculation pass. This improves running time at the cost of more memory usage.

Before:

```
> go test -run=XXX -bench=EnergySummation -count=8
BenchmarkEnergySummation-56    	      60	  19912731 ns/op
BenchmarkEnergySummation-56    	      58	  19881756 ns/op
BenchmarkEnergySummation-56    	      60	  20051701 ns/op
BenchmarkEnergySummation-56    	      57	  19925334 ns/op
BenchmarkEnergySummation-56    	      60	  19934364 ns/op
BenchmarkEnergySummation-56    	      60	  19991013 ns/op
BenchmarkEnergySummation-56    	      58	  20027357 ns/op
BenchmarkEnergySummation-56    	      58	  19939651 ns/op
```

After:

```
> go test -run=XXX -bench=EnergySummation -count=8
BenchmarkEnergySummation-56    	     222	   5395291 ns/op
BenchmarkEnergySummation-56    	     222	   5372478 ns/op
BenchmarkEnergySummation-56    	     222	   5375322 ns/op
BenchmarkEnergySummation-56    	     220	   5413824 ns/op
BenchmarkEnergySummation-56    	     220	   5358745 ns/op
BenchmarkEnergySummation-56    	     223	   5369265 ns/op
BenchmarkEnergySummation-56    	     223	   5369306 ns/op
BenchmarkEnergySummation-56    	     222	   5399840 ns/op
```